split referrers and location, reordered
