+++ src/shared/components/SectionWhere.js
@@ -30,20 +30,6 @@ function merge(name, data, comparatorData, comparatorLabel){
   return data;
 }
 
-function getReferrerUrls (data, i){
-  const maxLen = 60;
-  const displayString = data[0].length > maxLen ? data[0].substr(0, maxLen)+'…' : data[0];
-  let url = displayString.indexOf('http')<0 ? displayString : (
-    <a target="_blank" href={data[0]} >
-      {displayString}
-    </a>
-  );
-  return {
-    referrer: url,
-    views: data[1]
-  };
-}
-
 export default class SectionWhere extends React.Component {
 
   constructor(props) {
@@ -51,7 +37,7 @@ export default class SectionWhere extends React.Component {
   }
 
   render() {
-    if (!this.props.renderReferrers){
+    if (!this.props.renderWhere){
       return <div></div>
     }
 
@@ -59,9 +45,6 @@ export default class SectionWhere extends React.Component {
     let comparatorData = this.props.comparatorData;
     let countries = data.countries;
     let regions = data.regions.map((data) => mapTypes('region', data));
-    let refs = data.referrer_types.map((data) => mapTypes('referrer', data));
-    let socialrefs = data.social_referrers.sort().map((data) => mapTypes('referrer', data));
-    let refUrls = data.referrer_urls.map(getReferrerUrls);
     let keys = [articleLabel];
 
     if (comparatorData.regions){
@@ -74,7 +57,7 @@ export default class SectionWhere extends React.Component {
     return (<div>
       <Row>
         <Col xs={12}>
-          <h4>Where did the users come from?</h4>
+          <h4>Where in the world was the article read?</h4>
         </Col>
       </Row>
       <Row>
@@ -95,38 +78,6 @@ export default class SectionWhere extends React.Component {
         <Col xs={12} sm={6}>
           <Map
             data={countries}
-            />
-        </Col>
-      </Row>
-      <Row>
-        <Col xs={12}>
-          <h5>External Sources</h5>
-        </Col>
-      </Row>
-      <Row>
-        <Col xs={12} sm={6}>
-          <h6>Referrer Types</h6>
-          <BarChart
-            data={refs}
-            keys={['views']}
-            category={'referrer'}
-            yLabel="Page Views"
-            xLabel="Referrer"
-          />
-          <h6>Social Networks</h6>
-          <BarChart
-            data={socialrefs}
-            keys={['views']}
-            category={'referrer'}
-            yLabel="Page Views"
-            xLabel="Social Network"
-          />
-        </Col>
-        <Col xs={12} sm={6}>
-          <h6>Top URLs</h6>
-          <Table
-            headers={['Referrer', 'Views']}
-            rows={refUrls}
             />
         </Col>
       </Row>
