+++ src/shared/config/production.js
@@ -15,6 +15,7 @@ export default {
     'article:locations' : true,
     'article:bounceRate' : true,
     'article:who' : true,
+    'article:where' : true,
     'article:modifier:filters:UserCohort' : true,
     'article:modifier:filters:Referrers' : true,
     'article:modifier:filters:Region' : true,
